PairGenerator/requirements.txt
stanza~=1.10.1
nltk~=3.9.1
google_generativeai==0.8.5
python-decouple==3.8
protobuf~=5.29.4